pyproject.toml
[project]
name = "marketing-agent"
version = "0.1.0"
description = "eToro Marketing Agent - AI-powered creative generation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "anthropic>=0.18.0",
    "databricks-sdk>=0.15.0",
    "databricks-sql-connector>=3.0.0",
    "delta-spark>=3.0.0",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "openai>=1.3.0",  # For embeddings if needed
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
    "structlog>=23.2.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
